backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
pydantic[email]>=2.7.0,<3.0.0
psycopg[binary,pool]==3.2.3
python-slugify==8.0.4
httpx==0.27.2
python-dotenv==1.0.1
aiosmtplib==3.0.2
email-validator==2.1.1
python-jose[cryptography]==3.3.0
geoip2==4.8.0
apscheduler==3.10.4
pywebpush==1.14.0
# Bulk import dependencies
openpyxl==3.1.2
xlrd==2.0.1
aiofiles==24.1.0
pillow==10.4.0
python-multipart==0.0.9

